render image scales in cachable form.
By not using the fieldname but using the .tag() method of the images
view, the primary field is used with a UID in the URL.
This enables caching of the image since the URL now is unique and will
change when the image changes.
plone.namedfile >= 2.0.8 is required for the caching to work.