Add support to render the title independently of the text.
A new option on the slider pane (introduced in "ftw.slider" 3.1.0) allows to render the title of the pane without the need of adding a slider pane text. Until now the title was only rendered when the pane had some text.

See https://github.com/4teamwork/ftw.slider/pull/71 for the implementation in "ftw.slider".